+++ database/migrations/2025_11_30_014338_create_leave_type_balance_per_periods_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('leave_type_id')->constrained('leave_types')->onDelete('cascade');
             $table->unsignedInteger('from_period');
             $table->unsignedInteger('to_period')->nullable();
-            $table->integer('balance');
+            $table->decimal('balance', 7, 1);
             $table->timestamps();
         });
     }
