+++ database/migrations/2025_11_30_034451_create_employee_leave_type_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
 
             //Override Leave type initial eligibility balance
             $table->boolean('override_balance_upon_eligibility')->default(false);
-            $table->integer('balance_upon_eligibility')->nullable();
+            $table->decimal('balance_upon_eligibility', 7, 1)->nullable();
             $table->timestamps();
         });
     }
